+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Restructured code
 - ES6 export for tree-shaking
 - Add Component : `VCenter`
+- Add Component : `AboveFold`
 
 ## 1.0.1 - 2017-07-28
 - Add Component : `Lorem`